+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## Unreleased
+
+### Added
+- Manage Create Poll Request Op Builder
+- Manage poll op builder
+- Manage vote op builder
+- integration test for voting
+
+### Updated
+- XDR revision according to https://github.com/tokend/xdr/pull/50
+- XDR version 3.1.0
+
 ## [1.3.1-x.2] - 2019-03-21
 ### Added
 - New "bravo" blob type